KAN-112 [REFACTOR] Improve Navigation Structure/Logic
  * Improve navigation structure/logic for the navigation via BottomNavBar.

    - Apply Nested-Navigation graph for increased scalability, encapsulation of related screens, and better BackStackEntry Management.
    - For the robust navigation handling for destinations accessed via BottomNavBar, Each BottomNavBar triggers the navigation to the corresponding NavGraph, with appropriate nav configuration (Save states for related instances, and restore it on later visit; Prevent multiple instantiation of same object per every screen visit)
    - Implement Custom-defined BottomNavigationState composable to perform chronological back-button action handling.
    - Refactor category-specific NavDestination into one, generalized NavDestination that accepts the category as a parameter.
