+++ app/src/main/java/com/doyoonkim/knutice/navigation/MainServiceNavGraph.kt
@@ -14,46 +14,27 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.doyoonkim.common.navigation.BookmarkInfo
 import com.doyoonkim.common.navigation.Destination
 import com.doyoonkim.common.navigation.NavRoutes
 import com.doyoonkim.common.navigation.NoticeDetail
 import com.doyoonkim.common.ui.TipCategory
 import com.doyoonkim.knutice.di.components.AppComponent
-import com.doyoonkim.knutice.di.components.DaggerCustomerServiceSceneComponent
+import com.doyoonkim.knutice.di.components.DaggerCarrelStatusSceneComponent
 import com.doyoonkim.knutice.di.components.DaggerHomeSceneComponent
-import com.doyoonkim.knutice.di.components.DaggerNoticeByMajorSceneComponent
-import com.doyoonkim.knutice.di.components.DaggerNoticeDetailSceneComponent
-import com.doyoonkim.knutice.di.components.DaggerNoticeInCategorySceneComponent
-import com.doyoonkim.knutice.di.components.DaggerNoticeSearchSceneComponent
-import com.doyoonkim.knutice.di.components.DaggerNotificationPreferencesSceneComponent
-import com.doyoonkim.knutice.di.components.DaggerSettingsSceneComponent
+import com.doyoonkim.main.campus.carrel.CarrelStatusScreen
+import com.doyoonkim.main.campus.meal.DiningMenuScreen
 import com.doyoonkim.main.home.HomeDashboard
-import com.doyoonkim.main.notice.NoticeByMajorScreen
-import com.doyoonkim.main.notice.NoticeDetailScreen
-import com.doyoonkim.main.notice.NoticeSearchScreen
-import com.doyoonkim.main.notice.NoticesInCategoryScreen
-import com.doyoonkim.main.preference.CustomerServiceScreen
-import com.doyoonkim.main.preference.NotificationPreferencesScreen
-import com.doyoonkim.main.preference.OssNoticeScreen
-import com.doyoonkim.main.preference.UserPreferenceScreen
 import com.doyoonkim.main.tip.TipDetailScreen
-import com.doyoonkim.main.viewmodel.CustomerServiceViewModel
+import com.doyoonkim.main.viewmodel.CarrelStatusViewModel
 import com.doyoonkim.main.viewmodel.HomeViewModel
-import com.doyoonkim.main.viewmodel.NoticeByMajorViewModel
-import com.doyoonkim.main.viewmodel.NoticeDetailViewModel
-import com.doyoonkim.main.viewmodel.NoticeSearchViewModel
-import com.doyoonkim.main.viewmodel.NoticesInCategoryViewModel
-import com.doyoonkim.main.viewmodel.NotificationPreferencesViewModel
-import com.doyoonkim.main.viewmodel.SettingsViewModel
 import com.doyoonkim.model.NoticeCategory
 
 fun NavGraphBuilder.mainServiceNavGraph(
     navController: NavController,
     appComponent: AppComponent,
     onNoticeDetailRequested: (NoticeDetail) -> Unit,
-    onBookmarkServiceRequested: (BookmarkInfo) -> Unit,
-    onExit: () -> Unit = {  }
+    onTabSwitches: (String) -> Unit,
+    onPopBottomNavHistory: () -> Unit = {  }
 ) {
 
     // ViewModels will be injected via ViewModelFactory
@@ -73,17 +54,20 @@ fun NavGraphBuilder.mainServiceNavGraph(
             viewModel = viewModel<HomeViewModel>(factory = sceneComponent.viewModelFactory()),
             onSettingsRequested = { navController.navigate(NavRoutes.Settings.route) },
             onGoBackAction = {
-                navController.popBackStack().also { if (!it) onExit() }
+                onPopBottomNavHistory()
             },
             onMoreNoticeRequested = { dest ->
-                navController.run {
-                    when (dest) {
-                        Destination.MORE_GENERAL -> navigate(NavRoutes.GeneralNotices.route)
-                        Destination.MORE_ACADEMIC -> navigate(NavRoutes.AcademicNotices.route)
-                        Destination.MORE_SCHOLARSHIP -> navigate(NavRoutes.ScholarshipNotices.route)
-                        Destination.MORE_EVENT -> navigate(NavRoutes.EventNotices.route)
-                        Destination.MORE_EMPLOYMENT -> navigate(NavRoutes.EmploymentNotices.route)
-                        else -> { /* DO NOTHING. */
+                with(NavRoutes.NoticeList) {
+                    // Will be revised in 1.7.2 with updated use case.
+                    navController.run {
+                        when (dest) {
+                            Destination.MORE_GENERAL -> navigate(createRoute(NoticeCategory.GENERAL_NEWS.name))
+                            Destination.MORE_ACADEMIC -> navigate(createRoute(NoticeCategory.ACADEMIC_NEWS.name))
+                            Destination.MORE_SCHOLARSHIP -> navigate(createRoute(NoticeCategory.SCHOLARSHIP_NEWS.name))
+                            Destination.MORE_EVENT -> navigate(createRoute(NoticeCategory.EVENT_NEWS.name))
+                            Destination.MORE_EMPLOYMENT -> navigate(createRoute(NoticeCategory.EMPLOYMENT_NEWS.name))
+                            else -> { /* DO NOTHING. */
+                            }
                         }
                     }
                 }
@@ -95,393 +79,11 @@ fun NavGraphBuilder.mainServiceNavGraph(
                 navController.navigate("tipDetail/${category.name}/${Uri.encode(url)}")
             },
             onMoreMajorNoticeRequested = {
-                navController.navigate(NavRoutes.MajorNotices.route)
+                // Tab Switches. Revised in 1.7.2
+                onTabSwitches(GraphRoute.NOTICE)
             },
             onCarrelStatusRequested = { navController.navigate(NavRoutes.CarrelStatus.route) },
             onDiningMenuRequested = { navController.navigate(NavRoutes.DiningMenu.route) }
-        )
-    }
-
-    composable(
-        route = NavRoutes.MajorNotices.route
-    ) {
-        val sceneComponent = remember(appComponent) {
-            DaggerNoticeByMajorSceneComponent.factory().create(
-                systemServices = appComponent,
-                networkProvider = appComponent,
-                localPreferenceProvider = appComponent
-            )
-        }
-
-        NoticeByMajorScreen(
-            modifier = Modifier,
-            viewModel = viewModel<NoticeByMajorViewModel>(factory = sceneComponent.getViewModelFactory()),
-            onGoBackRequested = { navController.popBackStack() },
-            onSettingRequested = { navController.navigate(NavRoutes.Settings.route) },
-            onNoticeDetailRequested = { id, url -> onNoticeDetailRequested(NoticeDetail(id, url)) }
-        )
-    }
-
-    composable(
-        route = NavRoutes.NoticeSearch.route
-    ) {
-        val sceneComponent = remember(appComponent) {
-            DaggerNoticeSearchSceneComponent.factory().create(
-                systemServices = appComponent,
-                networkProvider = appComponent,
-                localStorageProvider = appComponent
-            )
-        }
-
-        NoticeSearchScreen(
-            modifier = Modifier,
-            viewModel = viewModel<NoticeSearchViewModel>(factory = sceneComponent.viewModelFactory()),
-            onBackPressed = { navController.popBackStack() },
-            onNoticeSelected = { id, url ->
-                onNoticeDetailRequested(NoticeDetail(id, url))
-            },
-            onBookmarkSelected = { id, title, category ->
-                onBookmarkServiceRequested(
-                    BookmarkInfo(
-                        noticeId = id,
-                        noticeTitle = title,
-                        noticeInfo = category
-                    )
-                )
-            }
-        )
-    }
-
-    composable(
-        route = NavRoutes.GeneralNotices.route,
-        enterTransition = {
-            slideIntoContainer(
-                animationSpec = tween(300, easing = EaseIn),
-                towards = AnimatedContentTransitionScope.SlideDirection.Start
-            )
-        },
-        exitTransition = {
-            slideOutOfContainer(
-                animationSpec = tween(300, easing = EaseOut),
-                towards = AnimatedContentTransitionScope.SlideDirection.End
-            )
-        }
-    ) {
-        val sceneComponent = remember(appComponent) {
-            DaggerNoticeInCategorySceneComponent.factory().create(
-                systemServices = appComponent,
-                networkProvider = appComponent
-            )
-        }
-
-        NoticesInCategoryScreen(
-            modifier = Modifier,
-            category = NoticeCategory.GENERAL_NEWS,
-            viewModel = viewModel<NoticesInCategoryViewModel>(factory = sceneComponent.viewModelFactory()),
-            onBackButtonPressed = { navController.popBackStack() },
-            onNoticeSelected = { id, url ->
-                onNoticeDetailRequested(NoticeDetail(id, url))
-            }
-        )
-    }
-
-    composable(
-        route = NavRoutes.AcademicNotices.route,
-        enterTransition = {
-            slideIntoContainer(
-                animationSpec = tween(300, easing = EaseIn),
-                towards = AnimatedContentTransitionScope.SlideDirection.Start
-            )
-        },
-        exitTransition = {
-            slideOutOfContainer(
-                animationSpec = tween(300, easing = EaseOut),
-                towards = AnimatedContentTransitionScope.SlideDirection.End
-            )
-        }
-    ) {
-        val sceneComponent = remember(appComponent) {
-            DaggerNoticeInCategorySceneComponent.factory().create(
-                systemServices = appComponent,
-                networkProvider = appComponent
-            )
-        }
-
-        NoticesInCategoryScreen(
-            modifier = Modifier,
-            category = NoticeCategory.ACADEMIC_NEWS,
-            viewModel = viewModel<NoticesInCategoryViewModel>(factory = sceneComponent.viewModelFactory()),
-            onBackButtonPressed = { navController.popBackStack() },
-            onNoticeSelected = { id, url ->
-                onNoticeDetailRequested(NoticeDetail(id, url))
-            }
-        )
-    }
-
-    composable(
-        route = NavRoutes.ScholarshipNotices.route,
-        enterTransition = {
-            slideIntoContainer(
-                animationSpec = tween(300, easing = EaseIn),
-                towards = AnimatedContentTransitionScope.SlideDirection.Start
-            )
-        },
-        exitTransition = {
-            slideOutOfContainer(
-                animationSpec = tween(300, easing = EaseOut),
-                towards = AnimatedContentTransitionScope.SlideDirection.End
-            )
-        }
-    ) {
-        val sceneComponent = remember(appComponent) {
-            DaggerNoticeInCategorySceneComponent.factory().create(
-                systemServices = appComponent,
-                networkProvider = appComponent
-            )
-        }
-
-        NoticesInCategoryScreen(
-            modifier = Modifier,
-            category = NoticeCategory.SCHOLARSHIP_NEWS,
-            viewModel = viewModel<NoticesInCategoryViewModel>(factory = sceneComponent.viewModelFactory()),
-            onBackButtonPressed = { navController.popBackStack() },
-            onNoticeSelected = { id, url ->
-                onNoticeDetailRequested(NoticeDetail(id, url))
-            }
-        )
-    }
-
-    composable(
-        route = NavRoutes.EventNotices.route,
-        enterTransition = {
-            slideIntoContainer(
-                animationSpec = tween(300, easing = EaseIn),
-                towards = AnimatedContentTransitionScope.SlideDirection.Start
-            )
-        },
-        exitTransition = {
-            slideOutOfContainer(
-                animationSpec = tween(300, easing = EaseOut),
-                towards = AnimatedContentTransitionScope.SlideDirection.End
-            )
-        }
-    ) {
-        val sceneComponent = remember(appComponent) {
-            DaggerNoticeInCategorySceneComponent.factory().create(
-                systemServices = appComponent,
-                networkProvider = appComponent
-            )
-        }
-
-        NoticesInCategoryScreen(
-            modifier = Modifier,
-            category = NoticeCategory.EVENT_NEWS,
-            viewModel = viewModel<NoticesInCategoryViewModel>(factory = sceneComponent.viewModelFactory()),
-            onBackButtonPressed = { navController.popBackStack() },
-            onNoticeSelected = { id, url ->
-                onNoticeDetailRequested(NoticeDetail(id, url))
-            }
-        )
-    }
-
-    composable(
-        route = NavRoutes.EmploymentNotices.route,
-        enterTransition = {
-            slideIntoContainer(
-                animationSpec = tween(300, easing = EaseIn),
-                towards = AnimatedContentTransitionScope.SlideDirection.Start
-            )
-        },
-        exitTransition = {
-            slideOutOfContainer(
-                animationSpec = tween(300, easing = EaseOut),
-                towards = AnimatedContentTransitionScope.SlideDirection.End
-            )
-        }
-    ) {
-        val sceneComponent = remember(appComponent) {
-            DaggerNoticeInCategorySceneComponent.factory().create(
-                systemServices = appComponent,
-                networkProvider = appComponent
-            )
-        }
-
-        NoticesInCategoryScreen(
-            modifier = Modifier,
-            category = NoticeCategory.EMPLOYMENT_NEWS,
-            viewModel = viewModel<NoticesInCategoryViewModel>(factory = sceneComponent.viewModelFactory()),
-            onBackButtonPressed = { navController.popBackStack() },
-            onNoticeSelected = { id, url ->
-                onNoticeDetailRequested(NoticeDetail(id, url))
-            }
-        )
-    }
-
-    // preferences
-    composable(
-        route = NavRoutes.Settings.route,
-        enterTransition = {
-            slideIntoContainer(
-                animationSpec = tween(300, easing = EaseIn),
-                towards = AnimatedContentTransitionScope.SlideDirection.Start
-            )
-        },
-        exitTransition = {
-            slideOutOfContainer(
-                animationSpec = tween(300, easing = EaseOut),
-                towards = AnimatedContentTransitionScope.SlideDirection.End
-            )
-        }
-    ) {
-        val sceneComponent = remember(appComponent) {
-            DaggerSettingsSceneComponent.factory().create(
-                systemServices = appComponent,
-                networkProvider = appComponent,
-                localStorageProvider = appComponent,
-                localPreferenceProvider = appComponent
-            )
-        }
-
-        UserPreferenceScreen(
-            modifier = Modifier.padding(horizontal = 10.dp),
-            viewModel = viewModel<SettingsViewModel>(factory = sceneComponent.viewModelFactory()),
-            onNotificationPreferenceClicked = { navController.navigate(NavRoutes.NotificationPreferences.route) },
-            onCustomerServiceClicked = { navController.navigate(NavRoutes.CustomerService.route) },
-            onOssClicked = { navController.navigate(NavRoutes.OpenSource.route) },
-            onBackPressed = { syncPerformed ->
-                if (syncPerformed) {
-                    navController.popBackStack(NavRoutes.Home.route, inclusive = true)
-                    navController.navigate(NavRoutes.Bookmark.route)
-                }
-                else navController.popBackStack()
-            }
-        )
-    }
-
-    composable(
-        route = NavRoutes.NotificationPreferences.route,
-        enterTransition = {
-            slideIntoContainer(
-                animationSpec = tween(300, easing = EaseIn),
-                towards = AnimatedContentTransitionScope.SlideDirection.Start
-            )
-        },
-        exitTransition = {
-            slideOutOfContainer(
-                animationSpec = tween(300, easing = EaseOut),
-                towards = AnimatedContentTransitionScope.SlideDirection.End
-            )
-        }
-    ) {
-        val sceneComponent = remember(appComponent) {
-            DaggerNotificationPreferencesSceneComponent.factory().create(
-                systemServices = appComponent,
-                localPreferenceProvider = appComponent,
-                networkProvider = appComponent
-            )
-        }
-
-        NotificationPreferencesScreen(
-            modifier = Modifier.padding(horizontal = 10.dp),
-            viewModel = viewModel<NotificationPreferencesViewModel>(factory = sceneComponent.viewModelFactory()),
-            onBackPressed = { navController.popBackStack() }
-        )
-    }
-
-    composable(
-        route = NavRoutes.CustomerService.route,
-        enterTransition = {
-            slideIntoContainer(
-                animationSpec = tween(300, easing = EaseIn),
-                towards = AnimatedContentTransitionScope.SlideDirection.Start
-            )
-        },
-        exitTransition = {
-            slideOutOfContainer(
-                animationSpec = tween(300, easing = EaseOut),
-                towards = AnimatedContentTransitionScope.SlideDirection.End
-            )
-        }
-    ) {
-        val sceneComponent = remember(appComponent) {
-            DaggerCustomerServiceSceneComponent.factory().create(
-                systemService = appComponent,
-                localPreferenceProvider = appComponent,
-                networkProvider = appComponent
-            )
-        }
-
-        CustomerServiceScreen(
-            modifier = Modifier.padding(horizontal = 10.dp),
-            viewModel = viewModel<CustomerServiceViewModel>(factory = sceneComponent.viewModelFactory()),
-            onBackPressed = { navController.popBackStack() }
-        )
-    }
-
-    composable(
-        route = NavRoutes.OpenSource.route,
-        enterTransition = {
-            slideIntoContainer(
-                animationSpec = tween(300, easing = EaseIn),
-                towards = AnimatedContentTransitionScope.SlideDirection.Start
-            )
-        },
-        exitTransition = {
-            slideOutOfContainer(
-                animationSpec = tween(300, easing = EaseOut),
-                towards = AnimatedContentTransitionScope.SlideDirection.End
-            )
-        }
-    ) {
-        OssNoticeScreen(
-            modifier = Modifier.padding(horizontal = 10.dp),
-            onBackPressed = { navController.popBackStack() }
-        )
-    }
-
-    // DeepLinks for NoticeDetailScreen
-    // notice?nttId=1234&contentUrl=https://ut.ac.kr/...&FabVisible=true
-    composable(
-        route = "notice?nttId={nttId}&contentUrl={contentUrl}&FabVisible={isFabVisible}",
-        enterTransition = {
-            slideIntoContainer(
-                animationSpec = tween(300, easing = EaseIn),
-                towards = AnimatedContentTransitionScope.SlideDirection.Start
-            )
-        },
-        exitTransition = {
-            slideOutOfContainer(
-                animationSpec = tween(300, easing = EaseOut),
-                towards = AnimatedContentTransitionScope.SlideDirection.End
-            )
-        }
-    ) { backStackEntry ->
-        val noticeInfo = backStackEntry.arguments?.let {
-            Triple(
-                it.getString("nttId")?.toInt() ?: 0,
-                Uri.decode(it.getString("contentUrl") ?: ""),
-                it.getString("isFabVisible").toBoolean() ?: false
-            )
-        } ?: Triple(0, "", false)
-
-        val sceneComponent = remember(appComponent) {
-            DaggerNoticeDetailSceneComponent.factory().create(
-                systemServices = appComponent,
-                networkProvider = appComponent,
-                firebaseInfrastructureProvider = appComponent
-            )
-        }
-
-        NoticeDetailScreen(
-            modifier = Modifier.fillMaxSize(),
-            viewModel = viewModel<NoticeDetailViewModel>(factory = sceneComponent.viewModelFactory()),
-            noticeInfo = noticeInfo,
-            onBookmarkCreate = { onBookmarkServiceRequested(BookmarkInfo(
-                noticeId = it.nttId,
-                noticeTitle = it.title,
-                noticeInfo = it.noticeName.ifBlank { it.departName }
-            )) },
-            onBackPressed = { navController.popBackStack() }
         )
     }
 
@@ -511,6 +113,69 @@ fun NavGraphBuilder.mainServiceNavGraph(
             modifier = Modifier.fillMaxSize(),
             tipCategory = tipDetail.first,
             contentUrl = tipDetail.second
+        ) {
+            navController.popBackStack()
+        }
+    }
+
+    // reading-room?roomId=ROOM3&seat=531
+    composable(
+        route = "reading-room?roomId={roomId}&seat={seatNo}",
+        enterTransition = {
+            slideIntoContainer(
+                animationSpec = tween(300, easing = EaseIn),
+                towards = AnimatedContentTransitionScope.SlideDirection.Start
+            )
+        },
+        exitTransition = {
+            slideOutOfContainer(
+                animationSpec = tween(300, easing = EaseOut),
+                towards = AnimatedContentTransitionScope.SlideDirection.End
+            )
+        }
+    ) { backStackEntry ->
+        val room = backStackEntry.arguments?.getString("roomId") ?: ""
+        val seat = backStackEntry.arguments?.getString("seatNo") ?: ""
+
+        val sceneComponent = remember(appComponent) {
+            DaggerCarrelStatusSceneComponent.factory().create(
+                systemServices = appComponent,
+                networkProvider = appComponent,
+                localCacheProvider = appComponent,
+                localPreferenceProvider = appComponent
+            )
+        }
+
+        CarrelStatusScreen(
+            modifier = Modifier,
+            viewModel = viewModel<CarrelStatusViewModel>(factory = sceneComponent.viewModelFactory()),
+            roomId = room
+        ) {
+            navController.popBackStack()
+        }
+    }
+
+    // meal?topic=STUDENT_CAFETERIA
+    composable(
+        route = "meal?topic={mealTopic}",
+        enterTransition = {
+            slideIntoContainer(
+                animationSpec = tween(300, easing = EaseIn),
+                towards = AnimatedContentTransitionScope.SlideDirection.Start
+            )
+        },
+        exitTransition = {
+            slideOutOfContainer(
+                animationSpec = tween(300, easing = EaseOut),
+                towards = AnimatedContentTransitionScope.SlideDirection.End
+            )
+        }
+    ) { backStackEntry ->
+        val selection = backStackEntry.arguments?.getString("mealTopic")
+
+        DiningMenuScreen(
+            modifier = Modifier,
+            hallSelection = selection ?: ""
         ) {
             navController.popBackStack()
         }
