+++ common/src/main/java/com/doyoonkim/common/navigation/NavRoutes.kt
@@ -11,20 +11,21 @@ sealed class NavRoutes(val route: String) {
     data object NotificationPreferences: NavRoutes(Destination.NOTIFICATION.name)
     data object CustomerService: NavRoutes(Destination.CS.name)
     data object OpenSource: NavRoutes(Destination.OSS.name)
-
-    data object GeneralNotices: NavRoutes(Destination.MORE_GENERAL.name)
-    data object AcademicNotices: NavRoutes(Destination.MORE_ACADEMIC.name)
-    data object ScholarshipNotices: NavRoutes(Destination.MORE_SCHOLARSHIP.name)
-    data object EventNotices: NavRoutes(Destination.MORE_EVENT.name)
-    data object EmploymentNotices: NavRoutes(Destination.MORE_EMPLOYMENT.name)
-
     data object NoticeSearch: NavRoutes(Destination.SEARCH.name)
 
     data object CarrelStatus: NavRoutes("reading-room")
     data object DiningMenu: NavRoutes("meal")
+
+    // Dynamic Notice List Destination
+    data object NoticeList: NavRoutes("notices_list/{category}") {
+        // Key value for access NavArgument
+        const val SELECTED_CATEGORY = "category"
+
+        fun createRoute(category: String) = "notices_list/$category"
+    }
 }
 
 enum class Destination {
     HOME, MORE_GENERAL, MORE_ACADEMIC, MORE_SCHOLARSHIP, MORE_EVENT, MORE_EMPLOYMENT,
-    SETTINGS, OSS, CS, SEARCH, NOTIFICATION, BOOKMARKS, MAJOR_NOTICES
+    SETTINGS, OSS, CS, SEARCH, NOTIFICATION, BOOKMARKS, MAJOR_NOTICES, NOTICES_LIST
 }
